Rename HeroSection Components
Renamed landing and reserve table HeroSection components given that the HeroComponent was renamed to better reflect what html element it returned.
