+++ src/Components/Landing/Main.js
@@ -1,13 +1,13 @@
 import React from "react";
 import AboutSection from "./AboutSection";
-import HeroSection from "./HeroSection";
 import HighlightsSection from "./HighlightsSection";
+import LandingHeroSection from "./LandingHeroSection";
 import TestimonialsSection from "./TestimonialsSection";
 
 export default function Main() {
   return (
     <main>
-      <HeroSection />
+      <LandingHeroSection />
       <HighlightsSection />
       <TestimonialsSection />
       <AboutSection />
